server/requirements.txt
python-dotenv
fastapi[all]
uvicorn
pipecat-ai[silero,websocket,google,openai]>=0.0.82
